Fix Boss Fight, update TopDownMovement, update SceneSwitchTrigger, update BoneProjectile, update enemy prefab sorting orders, update scenes, organize project folder
Boss fight is working - still things to adjust with tags ("Boss" vs "non_player"), and player cannot progress to finish point with wall blocking. Fix TopDownMovement.cs to find appropriate target after boss is destroyed. Update TopDownPlayer.prefab to use new health and weapon system. Update SceneSwitch.cs to now allow editors to specify a scene name to load otherwise it will load the next build scene. Fix BoneProjectile to check if it hits the player with the PlayerHealth component instaead of a tag with a player such as a bullet or bomb.