Refactor game manager, fix being able to pause during main menu
Refactor scripts to not find tag "ui_manager", and instead tell the game manager to update the ui. PlayerHealth and BossHealth still find "ui_manager" tag in awake. Does not update the player's health or the boss' health as there is only one player and one boss, and it is not necessary for the game object to track their health. Fix being able to turn on cross fade by pausing in the main menu.