Implement Scene persistence, allow correct player to be active depending on scene
Multiple changes to get scene persistence to work. Everything I've seen is stable. Objects that are not destroyed include the game manager, game ui canvas, main camera, and a new prefab called PlayerSideWithTop. PlayerSideWithTop is an empty object with both the sidescroller and top down player prefabs as separate children. SwitchPlayer.cs is stored on the PlayerSideWithTop prefab and controls which player to set active. Main menu now has its own script instead of using UIManager.cs. Scene cross fade transition is a prefab in each scene, and not as a child of the Game UI Canvas. Each scene's camera and object pools have been correctly initialized to test each scene independently without errors. SpawnPoint is a new prefab in every scene, that sets the location the player will appear when a new scene is loaded. Have to fix camera updating to correct position in rush scene. StickyPlatforms are currently not working because the "player" tags are children of the PlayerSideWithTop prefab.